+++ src/panels/RouterPanel.php
@@ -9,7 +9,9 @@ namespace yii\debug\panels;
 
 use Yii;
 use yii\base\InlineAction;
-use yii\debug\models\Router;
+use yii\debug\models\router\ActionRoutes;
+use yii\debug\models\router\CurrentRoute;
+use yii\debug\models\router\RouterRules;
 use yii\debug\Panel;
 use yii\log\Logger;
 
@@ -76,7 +78,11 @@ class RouterPanel extends Panel
      */
     public function getDetail()
     {
-        return Yii::$app->view->render('panels/router/detail', ['model' => new Router($this->data)]);
+        return Yii::$app->view->render('panels/router/detail', [
+            'currentRoute' => new CurrentRoute($this->data),
+            'routerRules' => new RouterRules(),
+            'actionRoutes' => new ActionRoutes(),
+        ]);
     }
 
     /**